feat: menambahkan status member jika login akan menunjukkan status active, dan pasif ketika logout

+++ app/Http/Controllers/AuthenticationController.php
@@ -16,21 +16,36 @@ class AuthenticationController extends Controller
             'email' => 'required|email',
             'password' => 'required',
         ]);
-
+    
         $user = Member::where('email', $request->email)->first();
-
+    
         if (! $user || ! Hash::check($request->password, $user->password)) {
             throw ValidationException::withMessages([
                 'email' => ['The provided credentials are incorrect.'],
             ]);
         }
-
+    
+        // Set status is_active menjadi true
+        $user->is_active = true;
+        $user->save();
+    
         return $user->createToken('user login')->plainTextToken;
     }
-
-
-    public function logout_member(Request $request){
-        $request->user()->currentAccessToken()->delete();
+    
+    public function logout_member(Request $request)
+    {
+        $user = $request->user();
+    
+        if ($user) {
+            // Set status is_active menjadi false
+            $user->is_active = false;
+            $user->save();
+    
+            // Hapus token akses saat ini
+            $request->user()->currentAccessToken()->delete();
+        }
+    
+        return response()->json(['message' => 'Logged out successfully'], 200);
     }
 
     public function me(Request $request){
